pyproject.toml
[project]
authors = [{name = "Johannes Tenhumberg", email = "johannes.tenhumberg@gmail.com"}]
name = "wzk"
requires-python = ">=3.13,<3.14"
version = "0.1.1"
dependencies = [
    "numpy>2.0",
    "scipy",
    "scikit-image",
    "scikit-learn",
    "einops",
    "msgpack>=1.1.2,<2",
    "ortools>=9.14.6206,<10",
    "matplotlib",
    "plotly",
    "yourdfpy>=0.0.59",
    "viser>=1.0.15,<2",
    "trimesh>=4.10.0,<5",
    "joblib",
    "fire",
    "meshcat>=0.3.2,<0.4",
    "jax>=0.9.0.1",
    "jaxtyping>=0.3.7",
]

[project.optional-dependencies]
wandb = [
    "wandb>=0.19",
]
gcp = [
    "google-cloud-compute>=1.22",
    "google-cloud-storage>=2.18",
]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "ty>=0.0.15",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
wzk = { path = ".", editable = true }

[tool.pixi.dependencies]
pixi-pycharm = ">=0.0.10,<0.0.11"
hatchling = ">=1.27.0,<2"


# Ruff


[tool.ty.src]
include = ["./", "src/", "tests/"]
exclude = [".venv/"]

[tool.ty.rules]
# Suppress false positives from numpy/matplotlib typing
not-subscriptable = "ignore"
invalid-argument-type = "ignore"
unsupported-operator = "ignore"
unresolved-attribute = "ignore"
no-matching-overload = "ignore"
invalid-parameter-default = "ignore"
not-iterable = "ignore"
invalid-assignment = "ignore"
invalid-return-type = "ignore"
call-non-callable = "ignore"
unknown-argument = "ignore"
missing-argument = "ignore"
invalid-method-override = "ignore"
positional-only-parameter-as-kwarg = "ignore"
invalid-declaration = "ignore"
possibly-missing-attribute = "ignore"
unused-type-ignore-comment = "warn"
# Optional dependencies
unresolved-import = "ignore"

[tool.ty.environment]
python-version = "3.13"


